# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LectureCreate": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "content_type": {
            "examples": [
              "video/mp4"
            ],
            "pattern": "^video/[\\w.+-]+$",
            "title": "Content Type",
            "type": "string"
          },
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "filename": {
            "examples": [
              "6.006-lecture-01.mp4"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "licence": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "CC BY-NC-SA 4.0"
            ],
            "title": "Licence"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "filename",
          "content_type"
        ],
        "title": "LectureCreate",
        "type": "object"
      },
      "LectureCreated": {
        "properties": {
          "lecture": {
            "$ref": "#/components/schemas/LectureOut"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadTarget"
          }
        },
        "required": [
          "lecture",
          "upload"
        ],
        "title": "LectureCreated",
        "type": "object"
      },
      "LectureOut": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "licence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licence"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/LectureStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        },
        "required": [
          "id",
          "title",
          "course_id",
          "status",
          "source_filename",
          "content_type",
          "size_bytes",
          "duration_s",
          "content_hash",
          "licence",
          "attribution",
          "visibility",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "LectureOut",
        "type": "object"
      },
      "LectureStatus": {
        "enum": [
          "awaiting_upload",
          "uploaded",
          "processing",
          "ready",
          "failed"
        ],
        "title": "LectureStatus",
        "type": "string"
      },
...
      "Progress": {
        "properties": {
          "done": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StageInfo"
            },
            "title": "Done",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "running": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Running",
            "type": "array"
          },
          "status": {
            "default": "running",
            "enum": [
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "Progress",
        "type": "object"
      },
      "ProgressEvent": {
        "description": "One server-sent event: the lecture's status and its latest run's progress.",
        "properties": {
          "lecture_status": {
            "$ref": "#/components/schemas/LectureStatus"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Progress"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "lecture_status",
          "run_id",
          "progress"
        ],
        "title": "ProgressEvent",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lecture_id": {
            "format": "uuid",
            "title": "Lecture Id",
            "type": "string"
          },
          "llm_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Usage"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageInfo"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "lecture_id",
          "status",
          "started_at",
          "finished_at",
          "error",
          "stages",
          "llm_usage"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "running",
          "succeeded",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "StageInfo": {
        "properties": {
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "seconds": {
            "title": "Seconds",
            "type": "number"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "seconds",
          "cached"
        ],
        "title": "StageInfo",
        "type": "object"
      },
...
      "UploadTarget": {
        "description": "Send the file with `method` to `url`, including `headers` exactly as given.",
        "properties": {
          "expires_in_s": {
            "title": "Expires In S",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "method": {
            "const": "PUT",
            "default": "PUT",
            "title": "Method",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "headers",
          "expires_in_s"
        ],
        "title": "UploadTarget",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Visibility": {
        "enum": [
          "public",
          "private"
        ],
        "title": "Visibility",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Lecture Summariser API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/lectures": {
      "get": {
        "description": "The lectures the caller may read: the public ones, and their own.",
        "operationId": "list_lectures_v1_lectures_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LectureOut"
                  },
                  "title": "Response List Lectures V1 Lectures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lectures",
        "tags": [
          "lectures"
        ]
      },
      "post": {
        "description": "A signed-in user's lecture is private to them. Without sign-in (local), it's public.",
        "operationId": "create_lecture_v1_lectures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LectureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Lecture",
        "tags": [
          "lectures"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/events": {
      "get": {
        "description": "Server-sent events: one JSON ProgressEvent per change, until the run finishes.",
        "operationId": "events_v1_lectures__lecture_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressEvent"
                }
              }
            },
            "description": "Server-sent events. Each `data:` line is a ProgressEvent as JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "processing"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/process": {
      "post": {
        "description": "Idempotent: while a run is in progress, asking again returns that run. Only the\nlecture's owner (or an admin) can start one.",
        "operationId": "process_v1_lectures__lecture_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process",
        "tags": [
          "processing"
        ]
      }
    },
    "/v1/lectures/{lecture_id}/runs": {
      "get": {
        "operationId": "list_runs_v1_lectures__lecture_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs V1 Lectures  Lecture Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "processing"
        ]
      }
    },
...
  }
}